+++ rest/settings.py
@@ -6,6 +6,7 @@ FLASK_PORT = 5343 # Hex representation of 'SC'
 REDIS_HOST = 'localhost'
 REDIS_PORT = 6379
 REDIS_DB = 0
+REDIS_PASSWORD = None
 
 # Kafka server information ------------
 KAFKA_HOSTS = ['localhost:9092']